+++ database/macrostrat/database/transfer/__init__.py
@@ -6,3 +6,11 @@ Requires Docker or locally-installed PostgreSQL client tools.
 from .dump_database import pg_dump, pg_dump_to_file
 from .move_tables import move_tables
 from .restore_database import pg_restore, pg_restore_from_file
+
+__all__ = [
+    "move_tables",
+    "pg_dump",
+    "pg_dump_to_file",
+    "pg_restore",
+    "pg_restore_from_file",
+]