+++ database/macrostrat/database/transfer/dump_database.py
@@ -1,5 +1,4 @@
 import asyncio
-import sys
 from pathlib import Path
 from typing import Optional
 
@@ -8,7 +7,7 @@ from sqlalchemy.engine import Engine
 
 from macrostrat.utils import get_logger
 
-from .stream_utils import print_stdout, print_stream_progress
+from .stream_utils import print_stderr, print_stdout, print_stream_progress
 from .utils import _create_command
 
 log = get_logger(__name__)
@@ -48,11 +47,11 @@ async def pg_dump(
 
 
 async def pg_dump_to_file(engine: Engine, dumpfile: Path | None, **kwargs):
-    proc = await pg_dump(engine, **kwargs)
-    if dumpfile is None or dumpfile == sys.stdout:
-        # If we have no dumpfile, just print to stdout
-        await _monitor_stdout(proc)
+    if dumpfile is None:
+        proc = await pg_dump(engine, stdout=None, **kwargs)
+        await _monitor_stderr(proc)
         return
+    proc = await pg_dump(engine, **kwargs)
     # Open dump file as an async stream
     async with aiofiles.open(dumpfile, mode="wb") as dest:
         await asyncio.gather(
@@ -61,8 +60,5 @@ async def pg_dump_to_file(engine: Engine, dumpfile: Path | None, **kwargs):
         )
 
 
-async def _monitor_stdout(proc):
-    await asyncio.gather(
-        asyncio.create_task(print_stdout(proc.stdout)),
-        asyncio.create_task(print_stream_progress(proc.stderr, None)),
-    )
+async def _monitor_stderr(proc):
+    await asyncio.gather(print_stderr(proc.stderr), proc.wait())